refactoring dataset and adding the dummy option to use mxrecord dataset
