+++ config.py
@@ -9,7 +9,9 @@ def get_args():
     parent_parser.add_argument('--data_root', type=str, default='')
     parent_parser.add_argument('--train_data_path', type=str, default='faces_emore/imgs')
     parent_parser.add_argument('--val_data_path', type=str, default='faces_emore')
+    parent_parser.add_argument('--use_mxrecord', action='store_true')
     parent_parser.add_argument('--train_data_subset', action='store_true')
+    parent_parser.add_argument('--swap_color_channel', action='store_true')
     parent_parser.add_argument('--prefix', type=str, default='default')
     parent_parser.add_argument('--gpus', type=int, default=1, help='how many gpus')
     parent_parser.add_argument('--distributed_backend', type=str, default='ddp', choices=('dp', 'ddp', 'ddp2'),)
