+++ dataset/image_folder_dataset.py
@@ -1,0 +1,64 @@
+import os
+
+import torchvision.datasets as datasets
+from PIL import Image
+import numpy as np
+import cv2
+
+from .augmenter import Augmenter
+
+class CustomImageFolderDataset(datasets.ImageFolder):
+
+    def __init__(self,
+                 root,
+                 transform=None,
+                 target_transform=None,
+                 loader=datasets.folder.default_loader,
+                 is_valid_file=None,
+                 low_res_augmentation_prob=0.0,
+                 crop_augmentation_prob=0.0,
+                 photometric_augmentation_prob=0.0,
+                 swap_color_channel=False,
+                 output_dir='./',
+                 ):
+
+        super(CustomImageFolderDataset, self).__init__(root,
+                                                       transform=transform,
+                                                       target_transform=target_transform,
+                                                       loader=loader,
+                                                       is_valid_file=is_valid_file)
+        self.root = root
+        self.augmenter = Augmenter(crop_augmentation_prob, photometric_augmentation_prob, low_res_augmentation_prob)
+        self.swap_color_channel = swap_color_channel
+        self.output_dir = output_dir  # for checking the sanity of input images
+
+    def __getitem__(self, index):
+        """
+        Args:
+            index (int): Index
+
+        Returns:
+            tuple: (sample, target) where target is class_index of the target class.
+        """
+        path, target = self.samples[index]
+        sample = self.loader(path)
+
+        if self.swap_color_channel:
+            # swap RGB to BGR if sample is in RGB
+            # we need sample in BGR
+            sample = Image.fromarray(np.asarray(sample)[:,:,::-1])
+
+        sample = self.augmenter.augment(sample)
+
+        sample_save_path = os.path.join(self.output_dir, 'training_samples', 'sample.jpg')
+        if not os.path.isfile(sample_save_path):
+            os.makedirs(os.path.dirname(sample_save_path), exist_ok=True)
+            cv2.imwrite(sample_save_path, np.array(sample))  # the result has to look okay (Not color swapped)
+
+        if self.transform is not None:
+            sample = self.transform(sample)
+        if self.target_transform is not None:
+            target = self.target_transform(target)
+
+        return sample, target
+
